refactor(db): split schema into pure Drizzle table files
- Rename schema/*.schema.ts → schema/*.table.ts (pure table defs, no Zod)
- Extract shared contentRenderingColumns into schema/_columns.ts
- Remove all drizzle-zod and zod imports from table definitions
