+++ packages/db/src/schema/_columns.ts
@@ -1,0 +1,6 @@
+import { integer, text } from 'drizzle-orm/pg-core';
+
+export const contentRenderingColumns = {
+  contentRendering: text('content_rendering'),
+  contentRenderingVersion: integer('content_rendering_version').notNull().default(1),
+};
