+++ packages/db/src/schema/article.table.ts
@@ -1,25 +1,7 @@
 import { relations } from 'drizzle-orm';
-import { type AnyPgColumn, index, pgTable } from 'drizzle-orm/pg-core';
-import { createInsertSchema, createUpdateSchema } from 'drizzle-zod';
-import { z } from 'zod/v4';
-import { validators } from '../lib/validation';
-import { user } from './auth.schema';
-
-/**
- * Relationship map:
- *
- * user ──< articles ──< comments ──< commentReactions
- *                  │         └──< comments (self-ref, replies)
- *                  └──< articleLikes
- *                  └──< articleViews
- *
- * articles   → author  : many-to-one  (articles.authorId → user.id)
- * articles   → comments: one-to-many
- * articles   → views   : one-to-many  (articleViews, one row per visit)
- * articles   → likes   : one-to-many  (articleLikes, keyed by hashed IP as visitorId)
- * comments   → parent  : self-ref     (comments.parentId → comments.id)
- * comments   → reactions: one-to-many (commentReactions)
- */
+import { type AnyPgColumn, index, pgTable, uniqueIndex } from 'drizzle-orm/pg-core';
+import { contentRenderingColumns } from './_columns';
+import { user } from './auth.table';
 
 export const articles = pgTable(
   'articles',
@@ -29,8 +11,7 @@ export const articles = pgTable(
     slug: t.varchar({ length: 255 }).notNull().unique(),
     description: t.varchar({ length: 255 }),
     content: t.text(),
-    contentRendering: t.text(),
-    contentRenderingVersion: t.integer().notNull().default(1),
+    ...contentRenderingColumns,
     imageUrl: t.varchar({ length: 255 }),
     isDraft: t.boolean().notNull().default(false),
     tags: t.text().array(),
@@ -39,7 +20,11 @@ export const articles = pgTable(
       .references(() => user.id, { onDelete: 'cascade' })
       .notNull(),
     createdAt: t.timestamp().defaultNow().notNull(),
-    updatedAt: t.timestamp({ mode: 'date', withTimezone: true }).$onUpdate(() => new Date()),
+    updatedAt: t
+      .timestamp({ mode: 'date', withTimezone: true })
+      .defaultNow()
+      .$onUpdate(() => new Date())
+      .notNull(),
   }),
   (t) => [
     index('articles_author_id_idx').on(t.authorId),
@@ -73,7 +58,11 @@ export const comments = pgTable(
     content: t.json().notNull(),
     parentId: t.uuid().references((): AnyPgColumn => comments.id, { onDelete: 'cascade' }),
     createdAt: t.timestamp().defaultNow().notNull(),
-    updatedAt: t.timestamp().defaultNow().notNull(),
+    updatedAt: t
+      .timestamp({ mode: 'date', withTimezone: true })
+      .defaultNow()
+      .$onUpdate(() => new Date())
+      .notNull(),
   }),
   (t) => [
     index('comments_article_id_idx').on(t.articleId),
@@ -100,6 +89,7 @@ export const commentReactions = pgTable(
   (t) => [
     index('comment_reactions_comment_id_idx').on(t.commentId),
     index('comment_reactions_user_id_idx').on(t.userId),
+    uniqueIndex('comment_reactions_comment_id_user_id_idx').on(t.commentId, t.userId),
   ],
 );
 
@@ -133,7 +123,7 @@ export const articleLikes = pgTable(
   }),
   (t) => [
     index('article_likes_article_id_idx').on(t.articleId),
-    index('article_likes_article_id_visitor_id_idx').on(t.articleId, t.visitorId),
+    uniqueIndex('article_likes_article_id_visitor_id_idx').on(t.articleId, t.visitorId),
   ],
 );
 
@@ -163,52 +153,3 @@ export const articleViewRelations = relations(articleViews, (t) => ({
     references: [articles.id],
   }),
 }));
-
-export const ArticleBaseSchema = z.object({
-  title: validators.title,
-  slug: validators.slug,
-  description: validators.description,
-  content: validators.content,
-  thumbnail: validators.thumbnail,
-  isDraft: validators.isDraft,
-  tags: validators.tags,
-});
-
-export const CreateArticleSchema = createInsertSchema(articles, {
-  title: ArticleBaseSchema.shape.title,
-  slug: ArticleBaseSchema.shape.slug,
-  description: ArticleBaseSchema.shape.description,
-  content: ArticleBaseSchema.shape.content,
-  isDraft: ArticleBaseSchema.shape.isDraft,
-  tags: ArticleBaseSchema.shape.tags,
-  authorId: z.string(),
-})
-  .omit({
-    id: true,
-    createdAt: true,
-    updatedAt: true,
-  })
-  .and(
-    z.object({
-      thumbnail: ArticleBaseSchema.shape.thumbnail,
-    }),
-  );
-
-export const UpdateArticleSchema = createUpdateSchema(articles, {
-  id: z.uuid(),
-  title: ArticleBaseSchema.shape.title,
-  slug: ArticleBaseSchema.shape.slug,
-  description: ArticleBaseSchema.shape.description,
-  content: ArticleBaseSchema.shape.content,
-  isDraft: ArticleBaseSchema.shape.isDraft,
-  tags: ArticleBaseSchema.shape.tags,
-})
-  .omit({
-    createdAt: true,
-    updatedAt: true,
-  })
-  .and(
-    z.object({
-      thumbnail: ArticleBaseSchema.shape.thumbnail,
-    }),
-  );
