+++ packages/db/src/schema/auth.table.ts
