+++ packages/db/src/schema/index.ts
@@ -1,8 +1,7 @@
-/** biome-ignore-all lint/performance/noBarrelFile: this is a barrel file */
-export * from './article.schema';
-export * from './auth.schema';
-export * from './experience.schema';
-export * from './guestbook.schema';
-export * from './project.schema';
-export * from './service.schema';
-export * from './snippet.schema';
+export * from './article.table';
+export * from './auth.table';
+export * from './experience.table';
+export * from './guestbook.table';
+export * from './project.table';
+export * from './service.table';
+export * from './snippet.table';
